try to see if CAII Works with zero config
